Complete TypeScript implementation (minus API call)
