+++ ts/package.json
@@ -25,6 +25,7 @@
     "@types/node-fetch": "^2.6.2",
     "jest": "^29.2.2",
     "jest-environment-jsdom": "^29.2.2",
+    "js-sha256": "^0.9.0",
     "next": "^12.3.3",
     "node-fetch": "^2.6.7",
     "parcel": "^2.7.0",
