+++ ts/src/registry/registry.test.ts
@@ -15,7 +15,7 @@
 */
 
 /*
-package pulldown
+package registry
 
 import (
 	"bufio"
